refactor: update twiml script to axios
